requirements.txt
selenium
beautifulsoup4
python-dateutil
pymongo
python-dotenv
streamlit
langchain 
chromadb 
llama-index 
torch 
transformers 
sentence-transformers 
llama-cpp-python
langchain-community
langchain-experimental
langchain-openai
langchain-core
langchain-anthropic
langchain-ollama
langchain-chroma
langgraph
langchain_mistralai
jq